@@ -1,33 +1,28 @@
+# frozen_string_literal: true
+
 source "https://rubygems.org"
 
 branch = ENV.fetch('SOLIDUS_BRANCH', 'master')
 gem 'solidus', git: 'https://github.com/solidusio/solidus.git', branch: branch
 gem 'solidus_auth_devise'
-gem 'deface', require: false
 
-if branch == 'master' || branch >= "v2.3"
-  gem 'rails', '~> 5.1.0' # hack for broken bundler dependency resolution
-elsif branch >= "v2.0"
-  gem 'rails', '~> 5.0.0' # hack for broken bundler dependency resolution
+# Needed to help Bundler figure out how to resolve dependencies,
+# otherwise it takes forever to resolve them
+if branch == 'master' || Gem::Version.new(branch[1..-1]) >= Gem::Version.new('2.10.0')
+  gem 'rails', '~> 6.0'
 else
-  gem "rails", "~> 4.2.7"
+  gem 'rails', '~> 5.0'
 end
 
-if branch < 'v2.5'
-  gem 'factory_bot', '5.1.1'
+case ENV['DB']
+when 'mysql'
+  gem 'mysql2'
+when 'postgresql'
+  gem 'pg'
 else
-  gem 'factory_bot', '5.1.1'
+  gem 'sqlite3'
 end
 
-if ENV['DB'] == 'mysql'
-  gem 'mysql2', '~> 0.4.10'
-else
-  gem 'pg', '~> 1.1'
-end
-
-group :development, :test do
-  gem "pry-rails"
-  gem 'selenium-webdriver'
-end
+gem 'solidus_extension_dev_tools', github: 'solidusio-contrib/solidus_extension_dev_tools'
 
 gemspec
