@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe "Orders", type: :feature, js: true do
@@ -11,7 +13,7 @@ describe "Orders", type: :feature, js: true do
 
     before do
       bundle.parts << [parts]
-      line_item.update_attributes!(quantity: 3)
+      line_item.update!(quantity: 3)
       order.reload.create_proposed_shipments
       order.finalize!
     end
@@ -115,7 +117,7 @@ describe "Orders", type: :feature, js: true do
           expect(alert_text).to include 'Please select the split destination'
           page.driver.browser.switch_to.alert.accept
           find('.select2-container').click
-          find(:xpath, '//body').find('.select2-drop li.select2-result', text: "#{location_name} (0 on hand)").click
+          find(:xpath, '//body').all('.select2-drop li.select2-result', text: "#{location_name} (0 on hand)")[0].click
           find('.save-split').click
         end
         expect(page).to have_content /Pending package from '#{location_name}'/i
