@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe "Parts", type: :feature, js: true do
   stub_authorization!
 
-  let!(:tshirt) { create(:product, :name => "T-Shirt") }
-  let!(:mug) { create(:product, :name => "Mug") }
+  let!(:tshirt) { create(:product, name: "T-Shirt") }
+  let!(:mug) { create(:product, name: "Mug") }
 
   before do
     visit spree.admin_product_path(mug)
