@@ -1,5 +1,6 @@
-Spree::Core::Engine.routes.draw do
+# frozen_string_literal: true
 
+Spree::Core::Engine.routes.draw do
   namespace :admin do
     resources :products do
       resources :parts do
@@ -15,5 +16,4 @@ Spree::Core::Engine.routes.draw do
       end
     end
   end
-
 end
